split `simple legacy classes` into a `legacy classes` one
The simple legacy classes migration isn't as simple anymore now that we
have to check the design system. Split it into 2 separate migrations
instead.
